+++ changelog.md
@@ -1,12 +1,23 @@
-### [0.0.2] - 2022-12-12
-#### Added
+### [0.0.3] - 2022-12-13
+
+#### Added  
+- Notification entitie
+- Notification content entitie
+- Send notification use case
+- Tests for notification and content
+- Tests for sendNotification useCase
+- In-memory database for tests
+
+### [0.0.2] - 2022-12-12  
+
+#### Added  
 - Basic Nest.js project
 - Prisma configuration
 - Prisma database
 - Validator for request parameters
   - Class to better typing the request parameters
 
-### [0.0.1] - 2022-12-12
+### [0.0.1] - 2022-12-12  
 
-#### Added
+#### Added  
 - Project creation